split FinishSignup in two so we can re-use SetPassword in the password recovery
