+++ unlock-app/src/__tests__/components/interface/FinishSignup.test.tsx
@@ -1,12 +1,10 @@
 import React from 'react'
 import * as rtl from 'react-testing-library'
 import {
-  passwordErrors,
   mapDispatchToProps,
-  validatePassword,
   FinishSignup,
-  Credentials,
 } from '../../../components/interface/FinishSignup'
+import { Credentials } from '../../../components/interface/SetPassword'
 
 let signupCredentials = jest.fn((credentials: Credentials) => credentials)
 
@@ -48,50 +46,6 @@ describe('FinishSignup', () => {
 
     // Also it should go to the loading spinner when you submit
     getByText('Creating Account...')
-  })
-
-  it('but not if the password does not match its confirmation', () => {
-    expect.assertions(1)
-
-    const emailAddress = 'geoff@bitconnect.gov'
-    const password = 'AVeryGoodPassword'
-
-    const { getByLabelText, getByDisplayValue } = rtl.render(
-      <FinishSignup
-        emailAddress={emailAddress}
-        signupCredentials={signupCredentials}
-      />
-    )
-
-    const inputs = [
-      getByLabelText('Password'),
-      getByLabelText('Confirm Password'),
-    ]
-    const submit = getByDisplayValue('Submit')
-
-    inputs.forEach((input, i) => {
-      rtl.fireEvent.change(input, { target: { value: password + i } })
-    })
-
-    rtl.fireEvent.click(submit)
-    expect(signupCredentials).not.toHaveBeenCalled()
-  })
-
-  describe('validatePassword', () => {
-    it('should not allow blank passwords', () => {
-      expect.assertions(1)
-      expect(validatePassword('', '')).toEqual([
-        passwordErrors.EMPTY,
-        passwordErrors.MID_LENGTH,
-      ])
-    })
-
-    it('should not allow passwords that are different from their confirmations', () => {
-      expect.assertions(1)
-      expect(
-        validatePassword('AVeryGoodPassword', 'PGeryVoodAassword')
-      ).toEqual([passwordErrors.NO_MATCH])
-    })
   })
 
   describe('mapDispatchToProps', () => {
